Do not require the port binding for the ssh service
Do not require the port binding for the ssh service to be on the 0.0.0.0 address.

The problem is that Windows Server 2016 docker implementation does not support "host IP addresses in NAT settings" - Removing the fixed IP solves the problem and makes no problems with *NIX based servers.

This fix was originally submitted to the docker-java project and has been merged, but was never published to the Jenkins plugin repository.