Explicitly set the SecurityProtocol when getting slave.jar
This explicitly sets allows the current powershell instance to accept TLS2, TLS1 and TLS connections.
This is crucial if Jenkins is hosted with https because otherwise the docker slaves cannot download the slave.jar from the server unless this security setting is set in the registry of the windows docker container.