Add a slave container configuration timeout
This should fix the following issue.
When a JobDSL seed job is executed while a container is waiting for its
configuration YAD plugin may loose track of the container.
The container is left behind indefinitely waiting for its configuration.

With this change, the container eventually exit in error and the jenkins
master is able to provision another container.